Take genet 5ae30cad and drop the parley patch rows
Genet 5ae30cad publishes the vendored Parley and Fontique forks as the
packages genet-parley and genet-fontique (lib names still parley and
fontique) and tightens the seam pins that publish exposed. Every genet.git
rev in this repo moves together, one rev for the family: the root manifest,
the genet_web_smoke example, and the graphshell web port. The six other
manifests that name genet.git do so only in comments; their dependencies
inherit through workspace = true and needed no edit.

The bump carries the fork publish plus these seam versions:
layout-dom-api 0.1.1, genet-scripted-dom 0.1.1, genet-static-dom 0.1.1,
genet-document-resources 0.2.1, genet-host-api 0.2.0, buckram 0.0.2,
genet-livery 0.0.3 and livery 0.0.4. The requirements this repo states by
exact version are aligned to match; the seam crates it names without a
version requirement need none.

The parley rows in both [patch.crates-io] tables are removed. Nothing in
genet answers to the name parley any more, so a row keyed on it matches no
package and becomes a resolution error rather than a silent no-op.
genet-livery reaches the fork through its own path dependency inside the
git checkout, so no row is needed to unify it. No fontique row existed.
The taffy (genet-taffy) and ipc-channel rows stay: those packages are
still named as they were.

So that document-canvas and content-contract do not fall back onto
registry parley, the workspace dependency now names the fork by package at
the same revision genet-livery is pinned to, keeping one genet-parley
identity in the graph; a registry copy beside the git one would double the
types.

genet-scripted-dom 0.1.1 changes NodeId::from_raw to take u64 where it took
usize, so the seven call sites in document-host's dom_view cast to u64
instead. The ids they carry are already u64 at every source, so this is the
cast changing, not the id types.

Resolving from git alone, mere at 5ae30cad and knot-editor still pinned at
7baa554c put two genet revs in one graph. That is transient and lives in no
committed lock, since mere does not commit its own: knot-editor is bumped
next with genet and mere moving together, and a one-line repin here follows.

Verified: mesquite 17 tests, cambium-genet-winit-host 80 (the current
count), document-host 3, and a check of every package that depends on a
genet crate. cambium-nematic's FeedEntry test literal is a known unrelated
failure, and document-host's caps test needs a prebuilt wasm guest that is
not present. No patch-not-used warning names parley or fontique.

+++ crates/script/document-host/src/dom_view.rs
@@ -35,7 +35,7 @@ pub fn snapshot(dom: &ScriptedDom, revision: u64) -> DocumentView {
 /// view if the id is not live.
 pub fn snapshot_subtree(dom: &ScriptedDom, revision: u64, root_id: u64) -> DocumentView {
     let mut nodes = Vec::new();
-    let root = NodeId::from_raw(root_id as usize);
+    let root = NodeId::from_raw(root_id as u64);
     if dom.is_live(root) {
         walk(dom, root, &mut nodes);
     }
@@ -93,7 +93,7 @@ pub fn apply(
             Mutation::AppendChild(a) => Some(a.parent),
         };
         if let Some(raw) = referenced {
-            if !dom.is_live(NodeId::from_raw(raw as usize)) {
+            if !dom.is_live(NodeId::from_raw(raw as u64)) {
                 return Err(TurnError::UnknownNode(raw));
             }
         }
@@ -111,20 +111,20 @@ pub fn apply(
 fn apply_one(dom: &mut ScriptedDom, m: Mutation) {
     match m {
         Mutation::SetText(SetTextArgs { node, text }) => {
-            dom.set_text(NodeId::from_raw(node as usize), &text);
+            dom.set_text(NodeId::from_raw(node as u64), &text);
         },
         Mutation::Remove(id) => {
-            dom.remove(NodeId::from_raw(id as usize));
+            dom.remove(NodeId::from_raw(id as u64));
         },
         Mutation::InsertBefore(InsertArgs { reference, new }) => {
-            let reference = NodeId::from_raw(reference as usize);
+            let reference = NodeId::from_raw(reference as u64);
             if let Some(parent) = dom.parent(reference) {
                 let child = create_block(dom, new);
                 dom.insert_before(parent, child, Some(reference));
             }
         },
         Mutation::AppendChild(AppendArgs { parent, new }) => {
-            let parent = NodeId::from_raw(parent as usize);
+            let parent = NodeId::from_raw(parent as u64);
             let child = create_block(dom, new);
             dom.append_child(parent, child);
         },
@@ -217,7 +217,7 @@ mod tests {
             },
         );
         assert!(matches!(r, Ok(1)), "set-text should apply: {r:?}");
-        assert_eq!(dom.text(NodeId::from_raw(t1_id as usize)), Some("Edited"));
+        assert_eq!(dom.text(NodeId::from_raw(t1_id as u64)), Some("Edited"));
 
         // append a <p> under body.
         let r = apply(
